修改异常
**复现步骤**

无

**原因和解决方案**

无

**影响范围**

无

**验证方案**

无

**涉及分支**

* main

**检查项**

- [ ] 需要更新依赖
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 编译通过
- [ ] 单元测试通过

/assign @songzhen

/cc @yuanchao @jinjie @zhuofeng

+++ server/controller/trisolaris/vtap/vtap.go
@@ -890,9 +890,9 @@ func (v *VTapInfo) updateCacheToDB() {
 			dbVTap.KernelVersion = cacheVTap.GetKernelVersion()
 			dbVTap.CtrlMac = cacheVTap.GetCtrlMac()
 			cacheExceptions := cacheVTap.GetExceptions()
-			tridentExceptions := int64(VTAP_TRIDENT_EXCEPTIONS_MASK) & cacheExceptions
-			controllerException := VTAP_CONTROLLER_EXCEPTIONS_MASK & dbVTap.Exceptions
-			dbVTap.Exceptions = controllerException | tridentExceptions
+			tridentExceptions := uint64(VTAP_TRIDENT_EXCEPTIONS_MASK) & uint64(cacheExceptions)
+			controllerException := uint64(VTAP_CONTROLLER_EXCEPTIONS_MASK) & uint64(dbVTap.Exceptions)
+			dbVTap.Exceptions = int64(controllerException | tridentExceptions)
 			cacheVTap.UpdateCurControllerIP(hostIP)
 			dbVTap.CurControllerIP = cacheVTap.GetCurControllerIP()
 			filterFlag = true
